backend/requirements.txt
alembic==1.10.3
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flask-vite==0.2.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
Pygments==2.15.0
python-dotenv==1.0.0
rich==12.6.0
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
